moved mock method
